I know how to use .self now
Items defined in the initializer for the Ability Box are now using .self so they are part of the object, and not just within the scope of the init.
Now I can call the items from outside the object, and attach listening functions them, that will be able to access the other parts of the frame.