MVC is in place with example
Here, the Character Name widget takes in any string. The Character class widget automatically updates along with it.
The controller now has a bidirectional relationship with the view, and the model.
If something in the view is updated, it lets the controller know, and it updates the model. If a widget wants to be notified of any field changes from the model, it can register with the controller. It must have a functions for 'update_field' defined.
The controller registers with the model for any fields that have been requested by widgets. If the model updates one of those fields, the controller gets notified, and lets each registered widget know.