MVC Fully Implemented
The model controls the data, and can speak to the controller.
The view controls the widgets, and can speak to the controller.
the controller owns the model and the view.
Widgets in the view can register with the controller, to be notified if a specific field is updated.
The model informs the controller of all field updates.
The controller receives a update from the model, it will inform the registered widgets in the view.