Empty ability score error removed
If something was entered into the ability score box, and then deleted leaving the box empty, an error was being thrown.
This had to do with the Model trying to do math on an empty string, as it couldn't convert it to an int.
A if statement was added to account for this. Now if the string is empty, it doesn't try to do math, and just places an empty string in turn.