View no longer dependent on DataObjects
All mentions of DataObjects in the views/frames have been removed. the view now calls to the controller to get this information. New methods have been added to controller to do this.
the controller and the model both still use the DataObjects, should eventually be replaced by a database call.