SQLite is now intergrated
CharacterData.db now has a API class so that the model can connect to it and retrieve data. It uses this currently to load a blank character into the character sheet. If the DB is changed, with different Character Stats, the model will grab that when the Character Page is opened.