Character Class and Race options
The data base has two new tables for character classes and character races, only two full options on each. The Class and Race widgets are comoboxes, that are drop down lists of the options provided in the database. all the columns from the DB are turned into attributes in Custom classes for each, and comma deliminated lists are parsed into arrays.