Name now updates by registering
The name field in the Character data object registers with the controller, to see if any widgets from the view update this field.
The widget calls the controller, informs it of the field and new value, the controller tell the registered object directly to update the field.
That field in the object also now calls the controller, to update the view.