Fields are registered now
Now when a widget needs to update a field, it does not call a related function from the controller to do so. All widgets call controller.trigger_field to make an update. The model is currently registering all fields in objects it holds with the controller for this to happen.
Currently the controller calls the related object directly, this should be moved to a model function. The controller shouldnt know about particular objects.
With this optimization, objects now own their own update functions, rather than the model having them for each object. I was able to remove a function from the controller and object for each case.