Character load improvements
setting class and race is now done on object level, rather than in the model.
Level can be updated from the UI now, and this updates the proficiency bonus. skills are all updated with new bonus.
Methods that were in DataObjects are now static methods of the Model. Now the controller and the model and the CharacterObject do not have to call this class. I think it can be removed now.